in_cpu: extend to support both interval_sec and interval_nsec for sub-second sample rate
10Hz: -i cpu -p interval_sec=0 -p interval_nsec=100000000
2Hz:  -i cpu -p interval_sec=0 -p interval_nsec=500000000
1Hz:  -i cpu -p interval_sec=1

+++ plugins/in_cpu/in_cpu.h
@@ -73,7 +73,7 @@ struct flb_in_cpu_config {
     int cpu_ticks;      /* CPU ticks (Kernel setting) */
     int coll_fd;        /* collector id/fd            */
     int interval_sec;   /* interval collection time (Second) */
-    long interval_nsec; /* interval collection time (Nanosecond) */
+    int interval_nsec;  /* interval collection time (Nanosecond) */
     struct cpu_stats cstats;
     struct flb_input_instance *i_ins;
 };
